+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.9.66
+
+* Some optional class features would not parse in as features.
+
 # 2.9.65
 
 * Feature effects: Planar Warrior, Ancestral Protectors, Sharpshooter, Crosbow Expert, and Savage Attacker provided by @Elwin#1410